Refactor: Move dashboard navItems to shared config to fix build
This resolves a potential build issue you were seeing. It was caused by importing navigation
configuration (navItems array and NavItem type) directly from
`src/app/dashboard/layout.tsx` into client-side page components.

Here's what I changed:
- I created a new file `src/config/navConfig.ts`.
- I moved the `NavItem` interface definition and the `navItems` array
  (including its lucide-react icon imports) from
  `src/app/dashboard/layout.tsx` into `src/config/navConfig.ts`.
- I updated `src/app/dashboard/layout.tsx`,
  `src/app/dashboard/user/page.tsx`,
  `src/app/dashboard/coach/page.tsx`, and
  `src/app/dashboard/admin/page.tsx` to import `navItems` and
  `NavItem` from the new `src/config/navConfig.ts` location.
- I cleaned up icon imports in `src/app/dashboard/layout.tsx`.

This centralization makes the navigation configuration more robustly
importable across different component types and is expected to resolve
the build failure.
